refactor(xrpl): time conversions
