+++ xrpl/time_test.go
@@ -1,4 +1,4 @@
-package utils
+package xrpl
 
 import (
 	"testing"
@@ -10,27 +10,27 @@ import (
 func TestTimeConversion_RippleTimeToUnixTime(t *testing.T) {
 	testCases := []struct {
 		name       string
-		rippleTime uint64
-		unixTime   uint64
+		rippleTime int64
+		unixTime   int64
 	}{
 		{
-			name:       "Ripple Time 0",
+			name:       "pass - ripple Time 0",
 			rippleTime: 0,
 			unixTime:   946684800000,
 		},
 		{
-			name:       "Ripple Time 1",
+			name:       "pass - ripple Time 1",
 			rippleTime: 1,
 			unixTime:   946684801000,
 		},
 		{
-			name:       "Ripple Time 100",
+			name:       "pass - ripple Time 100",
 			rippleTime: 100,
 			unixTime:   946684900000,
 		},
 	}
 	for _, tc := range testCases {
-		t.Run("Unix Time", func(t *testing.T) {
+		t.Run(tc.name, func(t *testing.T) {
 			actual := RippleTimeToUnixTime(tc.rippleTime)
 			assert.Equal(t, tc.unixTime, actual)
 		})
@@ -40,27 +40,27 @@ func TestTimeConversion_RippleTimeToUnixTime(t *testing.T) {
 func TestTimeConversion_UnixTimeToRippleTime(t *testing.T) {
 	testCases := []struct {
 		name       string
-		rippleTime uint64
-		unixTime   uint64
+		rippleTime int64
+		unixTime   int64
 	}{
 		{
-			name:       "Ripple Time 0",
+			name:       "pass - ripple Time 0",
 			rippleTime: 0,
 			unixTime:   946684800000,
 		},
 		{
-			name:       "Ripple Time 1",
+			name:       "pass - ripple Time 1",
 			rippleTime: 1,
 			unixTime:   946684801000,
 		},
 		{
-			name:       "Ripple Time 100",
+			name:       "pass - ripple Time 100",
 			rippleTime: 100,
 			unixTime:   946684900000,
 		},
 	}
 	for _, tc := range testCases {
-		t.Run("Unix Time", func(t *testing.T) {
+		t.Run(tc.name, func(t *testing.T) {
 			actual := UnixTimeToRippleTime(tc.unixTime)
 			assert.Equal(t, tc.rippleTime, actual)
 		})
@@ -70,7 +70,7 @@ func TestTimeConversion_UnixTimeToRippleTime(t *testing.T) {
 func TestTimeConversion_RippleTimeToIsoTime(t *testing.T) {
 	testCases := []struct {
 		name       string
-		rippleTime uint64
+		rippleTime int64
 		isoTime    string
 	}{
 		{
@@ -90,7 +90,7 @@ func TestTimeConversion_RippleTimeToIsoTime(t *testing.T) {
 		},
 	}
 	for _, tc := range testCases {
-		t.Run(tc.isoTime, func(t *testing.T) {
+		t.Run(tc.name, func(t *testing.T) {
 			actual := RippleTimeToISOTime(tc.rippleTime)
 			assert.Equal(t, tc.isoTime, actual)
 		})
@@ -100,7 +100,7 @@ func TestTimeConversion_RippleTimeToIsoTime(t *testing.T) {
 func TestTimeConversion_IsoTimeToRippleTime(t *testing.T) {
 	testCases := []struct {
 		name       string
-		rippleTime uint64
+		rippleTime int64
 		isoTime    string
 		wantErr    bool
 	}{
@@ -131,7 +131,7 @@ func TestTimeConversion_IsoTimeToRippleTime(t *testing.T) {
 	}
 
 	for _, tc := range testCases {
-		t.Run(tc.isoTime, func(t *testing.T) {
+		t.Run(tc.name, func(t *testing.T) {
 			actual, err := IsoTimeToRippleTime(tc.isoTime)
 			if tc.wantErr {
 				assert.Error(t, err)
@@ -142,36 +142,7 @@ func TestTimeConversion_IsoTimeToRippleTime(t *testing.T) {
 		})
 	}
 }
-func TestTimeConversion_FormatTimeToISO8601WithMillis(t *testing.T) {
-	testCases := []struct {
-		name     string
-		input    time.Time
-		expected string
-	}{
-		{
-			name:     "Time with milliseconds",
-			input:    time.Date(2023, 10, 1, 12, 34, 56, 789000000, time.UTC),
-			expected: "2023-10-01T12:34:56.789Z",
-		},
-		{
-			name:     "Time without milliseconds",
-			input:    time.Date(2023, 10, 1, 12, 34, 56, 0, time.UTC),
-			expected: "2023-10-01T12:34:56.000Z",
-		},
-		{
-			name:     "Time with zero milliseconds",
-			input:    time.Date(2000, 1, 1, 0, 0, 0, 0, time.UTC),
-			expected: "2000-01-01T00:00:00.000Z",
-		},
-	}
 
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			actual := formatTimeToISO8601WithMillis(tc.input)
-			assert.Equal(t, tc.expected, actual)
-		})
-	}
-}
 func TestTimeConversion_ParseISO8601(t *testing.T) {
 	testCases := []struct {
 		name     string
